+++ neoforge/kfflang/src/main/kotlin/thedarkcolour/kotlinforforge/neoforge/KotlinModLoadingContext.kt
@@ -15,7 +15,12 @@ public class KotlinModLoadingContext(private val container: KotlinModContainer) 
     public companion object {
         /** Mods should access through [MOD_CONTEXT] */
         public fun get(): KotlinModLoadingContext {
-            return ModLoadingContext.get().extension()
+            val activeContainer = ModLoadingContext.get().activeContainer
+            return if (activeContainer is KotlinModContainer) {
+                activeContainer.context
+            } else {
+                throw IllegalStateException("Tried to get KotlinModLoadingContext for active mod container, but ${activeContainer.modId} mod container was instance of ${activeContainer.javaClass.name}}")
+            }
         }
     }
 }